add update post functionlity

+++ backend/controllers/post.controller.js
@@ -57,12 +57,35 @@ export const getposts = async (req, res, next) => {
 }
 
 export const deletePost = async (req, res, next) => {
-    if (!req.user.isAdmin && req.user.id != req.params.userId) {
-        return next(errorHandler(403,"You are not allowed to delete this post"))
+    if (!req.user.isAdmin || req.user.id != req.params.userId) {
+        return next(errorHandler(403, "You are not allowed to delete this post"))
     }
     try {
         await Post.findByIdAndDelete(req.params.postId)
         res.status(200).json("This post has been deleted")
+    } catch (error) {
+        next(error)
+    }
+}
+
+export const updatePost = async (req, res, next) => {
+    if (!req.user.isAdmin || req.user.userId != req.params.userId) {
+        return next(errorHandler(403, "You are not allowed to update the post"))
+    }
+    try {
+        const updatedPost = await Post.findByIdAndUpdate(
+            req.params.postId,
+            {
+                $set: {
+                    title: req.body.title,
+                    content: req.body.content,
+                    category: req.body.category,
+                    image: req.body.image
+                }
+            },
+            { new: true }
+        )
+       return res.status(200).json(updatedPost)
     } catch (error) {
         next(error)
     }
